A really rough first display to look at reconstruction results.
This is a "not at all polished" event display that gives a first look
at the reconstruction.  It's based off of the superFGD event display,
but a lot of the features are copied yet, and it's not really
debugged.  For now, the user needs to adapt to the display, not the
other way around.