Add a class to save the trajectories
Trajectories are just a little bit to handy for physics analysis to
require loading the edepsim file everytime.  This lets most analysis
be done with just the CubeRecon file.  The hit segments and the
trajectories are saved.  The neutrino energy is not available in
edepsim, so that isn't available here.